backend/job-listener/requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.28.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-pubsub==2.18.1
google-cloud-resource-manager==1.10.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
idna==3.4
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
urllib3==1.26.16
